requirements/base.txt
playwright>=1.40.0
axe-playwright-python>=1.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
pyyaml>=6.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0